requirements.txt
altgraph==0.17.3
asgiref==3.3.4
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Cython==0.29.32
dj-database-url==0.5.0
Django==3.2.2
django-configurations==2.2
django-cors==0.1
django-cors-headers==3.13.0
django-debug-toolbar==3.2.1
django-extensions==3.1.3
djangorestframework==3.14.0
drf-yasg==1.21.4
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
macholib==1.16.2
MarkupSafe==2.1.1
packaging==21.3
psutil==5.9.2
psycopg2-binary==2.9.3
py-cpuinfo==8.0.0
pyinstaller==5.4.1
pyinstaller-hooks-contrib==2022.10
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2021.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==5.2.0
